requirements.txt
```
mysql-connector-python
cryptography
bcrypt
```
